feat(#3bngwaj): rename components and stories name
